blueprint: Directly use elasticsearch.deploy()
The old blueprint called deploy() on the Quilt deployment,
and passed in the elasticsearch application.  This added an
unnecessary level of indirection; we should call the applications'
deploy() method directly instead.

This commit also changes Elasticsearch's deploy() function
to directly call the deploy() method on its containers and
load balancer.